User: Add basic support for email verification :)
When you sign up with email or change it, the 128-byte token (encoded
with base64) is generated and an e-mail with link is sent to given
address.

Additionally, this fixes an issue with checking username availabity
that regressed after making profiles private.

You need an SMTP server installed for it to work.