Tilify: Rework a way Data Tables work
The columns are now entirely defined by user, which get the whole
entry (not only a single property) to generate it.

This commit adds also a separate type of "column" - "control". This is
basically a button (only them are allowed from widgets for now) which
calls a callback with a corresponding entry when clicked.